refactor tests for $$.sizeLessThanOrEqual
replaced useless unit-tests by analogous integration test
